# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PakDeviceKind": {
        "enum": [
          "engineering",
          "otk_line"
        ],
        "title": "PakDeviceKind",
        "type": "string"
      },
...
      "PakDeviceSummaryResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PakDeviceKind"
          }
        },
        "required": [
          "id",
          "code",
          "kind"
        ],
        "title": "PakDeviceSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationSessionDetailResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "firmware_version": {
            "title": "Firmware Version",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kg_dev_eui": {
            "title": "Kg Dev Eui",
            "type": "string"
          },
          "last_activity_at": {
            "format": "date-time",
            "title": "Last Activity At",
            "type": "string"
          },
          "pak": {
            "$ref": "#/components/schemas/PakDeviceSummaryResponse"
          },
          "pak_id": {
            "format": "uuid",
            "title": "Pak Id",
            "type": "string"
          },
          "slot_no": {
            "title": "Slot No",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/VerificationSessionStatus"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/VerificationStepResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "total_steps": {
            "title": "Total Steps",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kg_dev_eui",
          "pak_id",
          "pak",
          "slot_no",
          "firmware_version",
          "total_steps",
          "status",
          "started_at",
          "last_activity_at",
          "completed_at",
          "created_at",
          "updated_at",
          "steps"
        ],
        "title": "VerificationSessionDetailResponse",
        "type": "object"
      },
      "VerificationSessionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VerificationSessionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "VerificationSessionListResponse",
        "type": "object"
      },
      "VerificationSessionResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "firmware_version": {
            "title": "Firmware Version",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kg_dev_eui": {
            "title": "Kg Dev Eui",
            "type": "string"
          },
          "last_activity_at": {
            "format": "date-time",
            "title": "Last Activity At",
            "type": "string"
          },
          "pak": {
            "$ref": "#/components/schemas/PakDeviceSummaryResponse"
          },
          "pak_id": {
            "format": "uuid",
            "title": "Pak Id",
            "type": "string"
          },
          "slot_no": {
            "title": "Slot No",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/VerificationSessionStatus"
          },
          "total_steps": {
            "title": "Total Steps",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kg_dev_eui",
          "pak_id",
          "pak",
          "slot_no",
          "firmware_version",
          "total_steps",
          "status",
          "started_at",
          "last_activity_at",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "VerificationSessionResponse",
        "type": "object"
      },
      "VerificationSessionStatus": {
        "enum": [
          "RUNNING",
          "PASSED",
          "FAILED",
          "ABORTED",
          "INCOMPLETE"
        ],
        "title": "VerificationSessionStatus",
        "type": "string"
      },
      "VerificationStepResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "defect_group_id": {
            "format": "uuid",
            "title": "Defect Group Id",
            "type": "string"
          },
          "error_group_code": {
            "title": "Error Group Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "measurement_max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Max Value"
          },
          "measurement_min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Min Value"
          },
          "measurement_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Unit"
          },
          "measurement_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Value"
          },
          "pak_test_id": {
            "format": "uuid",
            "title": "Pak Test Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/VerificationStepStatus"
          },
          "step_no": {
            "title": "Step No",
            "type": "integer"
          },
          "test_label": {
            "title": "Test Label",
            "type": "string"
          },
          "test_name": {
            "title": "Test Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "step_no",
          "pak_test_id",
          "defect_group_id",
          "test_name",
          "test_label",
          "error_group_code",
          "status",
          "measurement_value",
          "measurement_min_value",
          "measurement_max_value",
          "measurement_unit",
          "started_at",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "VerificationStepResponse",
        "type": "object"
      },
      "VerificationStepStatus": {
        "enum": [
          "RUNNING",
          "PASSED",
          "FAILED",
          "ABORTED"
        ],
        "title": "VerificationStepStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "backend",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/verification/sessions": {
      "get": {
        "operationId": "verification-list_sessions",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "pak_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pak Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VerificationSessionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "started_at",
              "enum": [
                "kg_dev_eui",
                "status",
                "started_at",
                "completed_at",
                "created_at",
                "updated_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSessionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "verification"
        ]
      }
    },
    "/verification/sessions/{session_id}": {
      "get": {
        "operationId": "verification-get_session",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSessionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "verification"
        ]
      }
    }
  }
}